python/requirements.txt
black==23.7.0
boto3==1.28.16
botocore==1.31.16
click==8.1.6
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
